method that accepts parameter of trait
