+++ traits/src/impls/params.rs
@@ -1,0 +1,5 @@
+use crate::sumarizer::summary::Summary;
+
+pub fn broadcast(t: &impl Summary) {
+    println!("Breaking News: {}", t.summarize());
+}
