+++ traits/src/lib.rs
@@ -1,1 +1,2 @@
 pub mod sumarizer;
+pub mod impls;
